introduced 'models' and 'transducers' packages
