+++ test/test_beam_search.py
@@ -1,6 +1,5 @@
 import unittest
 
-import dynet_config
 import dynet as dy
 
 from xnmt.attenders import MlpAttender
@@ -9,12 +8,12 @@ from xnmt.bridges import CopyBridge
 from xnmt.decoders import AutoRegressiveDecoder
 from xnmt.embedders import SimpleWordEmbedder
 from xnmt.input_readers import PlainTextReader
-from xnmt.recurrent_transducers import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from transducers.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
 from xnmt.loss_calculators import AutoRegressiveMLELoss
 from xnmt.param_collections import ParamManager
 from xnmt.transforms import NonLinear
 from xnmt.scorers import Softmax
-from xnmt.translators import DefaultTranslator
+from models.translators import DefaultTranslator
 from xnmt.search_strategies import BeamSearch, GreedySearch
 from xnmt.vocabs import Vocab
