+++ test/test_encoders.py
@@ -2,7 +2,6 @@ import unittest
 import math
 
 import numpy as np
-import dynet_config
 import dynet as dy
 
 from xnmt.attenders import MlpAttender
@@ -10,13 +9,13 @@ from xnmt.bridges import CopyBridge
 from xnmt.decoders import AutoRegressiveDecoder
 from xnmt.embedders import SimpleWordEmbedder
 from xnmt.input_readers import PlainTextReader
-from xnmt.recurrent_transducers import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from transducers.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
 from xnmt.param_collections import ParamManager
-from xnmt.pyramidal import PyramidalLSTMSeqTransducer
+from transducers.pyramidal import PyramidalLSTMSeqTransducer
 from xnmt.scorers import Softmax
-from xnmt.self_attention import MultiHeadAttentionSeqTransducer
+from transducers.self_attention import MultiHeadAttentionSeqTransducer
 from xnmt.transforms import NonLinear
-from xnmt.translators import DefaultTranslator
+from models.translators import DefaultTranslator
 from xnmt.vocabs import Vocab
 from xnmt import batchers, events
