+++ xnmt/batchers.py
@@ -1,16 +1,16 @@
 import warnings
-from typing import Any, Callable, Optional, Sequence, Tuple, Union
+from typing import Callable, Optional, Sequence, Tuple, Union
 import math
 import random
 from abc import ABC, abstractmethod
+from functools import lru_cache
 
 import numpy as np
 import dynet as dy
 
-from xnmt.vocabs import Vocab
 from xnmt.persistence import serializable_init, Serializable
-from xnmt import expression_seqs, recurrent_transducers
-from functools import lru_cache
+from xnmt import expression_seqs
+from xnmt.transducers import recurrent
 from xnmt import sent
 
 class Batch(ABC):
@@ -654,8 +654,8 @@ class WordTrgSrcBatcher(WordSortBatcher, Serializable):
       self.batch_size = (sum([s.sent_len() for s in src]) + sum([s.sent_len() for s in trg])) / len(src) * self.avg_batch_size
     return super()._pack_by_order(src, trg, order)
 
-def truncate_batches(*xl: Union[dy.Expression, Batch, Mask, recurrent_transducers.UniLSTMState]) \
-        -> Sequence[Union[dy.Expression, Batch, Mask, recurrent_transducers.UniLSTMState]]:
+def truncate_batches(*xl: Union[dy.Expression, Batch, Mask, recurrent.UniLSTMState]) \
+        -> Sequence[Union[dy.Expression, Batch, Mask, recurrent.UniLSTMState]]:
   """
   Truncate a list of batched items so that all items have the batch size of the input with the smallest batch size.
 
@@ -678,7 +678,7 @@ def truncate_batches(*xl: Union[dy.Expression, Batch, Mask, recurrent_transducer
       batch_sizes.append(len(x))
     elif isinstance(x, Mask):
       batch_sizes.append(x.batch_size())
-    elif isinstance(x, recurrent_transducers.UniLSTMState):
+    elif isinstance(x, recurrent.UniLSTMState):
       batch_sizes.append(x.output().dim()[1])
     else:
       raise ValueError(f"unsupported type {type(x)}")
@@ -692,7 +692,7 @@ def truncate_batches(*xl: Union[dy.Expression, Batch, Mask, recurrent_transducer
         ret.append(mark_as_batch(x[:min(batch_sizes)]))
       elif isinstance(x, Mask):
         ret.append(Mask(x.np_arr[:min(batch_sizes)]))
-      elif isinstance(x, recurrent_transducers.UniLSTMState):
+      elif isinstance(x, recurrent.UniLSTMState):
         ret.append(x[:,:min(batch_sizes)])
       else:
         raise ValueError(f"unsupported type {type(x)}")
