+++ xnmt/eval_tasks.py
@@ -6,7 +6,7 @@ import dynet as dy
 
 from xnmt.batchers import Batcher
 from xnmt.eval_metrics import Evaluator, EvalScore
-from xnmt import inferences, input_readers, model_base
+from xnmt import inferences, input_readers
 from xnmt.persistence import serializable_init, Serializable, Ref, bare
 from xnmt.loss_calculators import LossCalculator, AutoRegressiveMLELoss
 from xnmt.eval_metrics import LossScore
