+++ xnmt/models/base.py
