+++ xnmt/specialized_encoders/segmenting_encoder/segmenting_encoder.py
@@ -9,11 +9,11 @@ from xnmt.batchers import Mask
 from xnmt.events import register_xnmt_handler, handle_xnmt_event
 from xnmt.expression_seqs import ExpressionSequence
 from xnmt.persistence import serializable_init, Serializable, Ref, bare
-from xnmt.transducers import SeqTransducer, FinalTransducerState, IdentitySeqTransducer
+from xnmt.transducers.base import SeqTransducer, FinalTransducerState, IdentitySeqTransducer
 from xnmt.losses import FactoredLossExpr
 from xnmt.specialized_encoders.segmenting_encoder.priors import GoldInputPrior
 from xnmt.reports import Reportable
-from xnmt.recurrent_transducers import BiLSTMSeqTransducer
+from xnmt.transducers.recurrent import BiLSTMSeqTransducer
 from xnmt.specialized_encoders.segmenting_encoder.segmenting_composer import SeqTransducerComposer
 from xnmt.expression_seqs import CompoundSeqExpression
