+++ xnmt/transforms.py
@@ -1,10 +1,7 @@
 import dynet as dy
-from typing import List
 
 from xnmt.param_collections import ParamManager
 from xnmt.param_initializers import GlorotInitializer, ZeroInitializer
-from xnmt.expression_seqs import ExpressionSequence
-from xnmt.transducers import SeqTransducer, FinalTransducerState
 from xnmt.persistence import serializable_init, Serializable, bare, Ref
 
 class Transform(object):
@@ -193,22 +190,3 @@ class MLP(Transform, Serializable):
       expr = layer(expr)
     return expr
 
-class TransformSeqTransducer(SeqTransducer, Serializable):
-  yaml_tag = '!TransformSeqTransducer'
-
-  @serializable_init
-  def __init__(self, transform: Transform):
-    """
-    Args:
-      transform: the Transform to apply to the sequence
-    """
-    self.transform = transform
-
-  def get_final_states(self) -> List[FinalTransducerState]:
-    return self._final_states
-
-  def transduce(self, src: ExpressionSequence) -> ExpressionSequence:
-    output = self.transform(src.as_tensor())
-    output_seq = ExpressionSequence(expr_tensor=output)
-    self._final_states = [FinalTransducerState(output_seq[-1])]
-    return output_seq
